redundancy check

+++ util/download_iEEG_data.py
@@ -45,6 +45,7 @@ python3.6 -c 'import get_iEEG_data; get_iEEG_data.get_iEEG_data("arevell", "pass
 with open(outputfile, 'rb') as f: data, fs = pickle.load(f)
 """
 import sys
+import os
 sys.path.append('../../ieegpy/ieeg')
 from ieeg.auth import Session
 import pandas as pd
@@ -57,34 +58,39 @@ def get_iEEG_data(username, password, iEEG_filename, start_time_usec, stop_time_
     print("start_time_usec: {0}".format(start_time_usec))
     print("stop_time_usec: {0}".format(stop_time_usec))
     print("ignore_electrodes: {0}".format(ignore_electrodes))
-    start_time_usec = int(start_time_usec)
-    stop_time_usec = int(stop_time_usec)
-    duration = stop_time_usec - start_time_usec
-    s = Session(username, password)
-    ds = s.open_dataset(iEEG_filename)
-    channels = list(range(len(ds.ch_labels)))
-    fs = ds.get_time_series_details(ds.ch_labels[0]).sample_rate #get sample rate
-    
-    #if duration is greater than ~10 minutes, then break up the API request to iEEG.org. 
-    #The server blocks large requests, so the below code breaks up the request and 
-    #concatenates the data
-    server_limit_minutes = 5
-    if duration < server_limit_minutes*60*1e6:
-        data = ds.get_data(start_time_usec, duration, channels)
-    if duration >= server_limit_minutes*60*1e6:
-        break_times = np.ceil(np.linspace(start_time_usec, stop_time_usec, num=int(np.ceil(duration/(server_limit_minutes*60*1e6))+1), endpoint=True))
-        break_data = np.zeros(shape = (int(np.ceil(duration/1e6*fs)), len(channels)))#initialize
-        print("breaking up data request from server because length is too long")
-        for i in range(len(break_times)-1):
-            print("{0}/{1}".format(i+1, len(break_times)-1))
-            break_data[range(int( np.ceil((break_times[i]-break_times[0])/1e6*fs) ), int(  np.ceil((break_times[i+1]- break_times[0])/1e6*fs) )  ),:] = ds.get_data(break_times[i], break_times[i+1]-break_times[i], channels)
-        data = break_data
-    df = pd.DataFrame(data, columns=ds.ch_labels)
-    df = pd.DataFrame.drop(df, ignore_electrodes, axis=1)
-    print("Saving to: {0}".format(outputfile))
-    with open(outputfile, 'wb') as f: pickle.dump([df, fs], f, protocol=4)
-    print("...done\n")
 
+    if not os.path.isfile(outputfile):
+        # data has not yet been downloaded for this interval
+        start_time_usec = int(start_time_usec)
+        stop_time_usec = int(stop_time_usec)
+        duration = stop_time_usec - start_time_usec
+        s = Session(username, password)
+        ds = s.open_dataset(iEEG_filename)
+        channels = list(range(len(ds.ch_labels)))
+        fs = ds.get_time_series_details(ds.ch_labels[0]).sample_rate #get sample rate
+        
+        #if duration is greater than ~10 minutes, then break up the API request to iEEG.org. 
+        #The server blocks large requests, so the below code breaks up the request and 
+        #concatenates the data
+        server_limit_minutes = 5
+        if duration < server_limit_minutes*60*1e6:
+            data = ds.get_data(start_time_usec, duration, channels)
+        if duration >= server_limit_minutes*60*1e6:
+            break_times = np.ceil(np.linspace(start_time_usec, stop_time_usec, num=int(np.ceil(duration/(server_limit_minutes*60*1e6))+1), endpoint=True))
+            break_data = np.zeros(shape = (int(np.ceil(duration/1e6*fs)), len(channels)))#initialize
+            print("breaking up data request from server because length is too long")
+            for i in range(len(break_times)-1):
+                print("{0}/{1}".format(i+1, len(break_times)-1))
+                break_data[range(int( np.ceil((break_times[i]-break_times[0])/1e6*fs) ), int(  np.ceil((break_times[i+1]- break_times[0])/1e6*fs) )  ),:] = ds.get_data(break_times[i], break_times[i+1]-break_times[i], channels)
+            data = break_data
+        df = pd.DataFrame(data, columns=ds.ch_labels)
+        df = pd.DataFrame.drop(df, ignore_electrodes, axis=1)
+        print("Saving to: {0}".format(outputfile))
+        with open(outputfile, 'wb') as f: pickle.dump([df, fs], f, protocol=4)
+        print("...done\n")
+    else:
+        # data has already been downloaded for this interval
+        print("{} exists, skipping...".format(outputfile))
 
 """"
 Download and install iEEG python package - ieegpy
